added local provider new ppkg struct
