+++ .github/workflows/testingAws.yml
@@ -80,8 +80,8 @@ jobs:
         working-directory: test
         if: ${{ matrix.os == 'windows-latest' }}
         run: |
-          # go test -bench=BenchmarkAwsTestingManaged -benchtime=1x -cover -v
-          go test -bench=BenchmarkAwsTestingHA -benchtime=1x -cover -v
+          # go test -tags testing_aws -bench=BenchmarkAwsTestingManaged -benchtime=1x -cover -v
+          go test -tags testing_aws -bench=BenchmarkAwsTestingHA -benchtime=1x -cover -v
 
       - name: integration testing
         if: ${{ matrix.os != 'windows-latest' }}
