+++ .github/workflows/testingAzure.yml
@@ -82,8 +82,8 @@ jobs:
         working-directory: test
         if: ${{ matrix.os == 'windows-latest' }}
         run: |
-          go test -bench=BenchmarkAzureTestingManaged -benchtime=1x -cover -v
-          go test -bench=BenchmarkAzureTestingHA -benchtime=1x -cover -v
+          go test -tags testing_azure -bench=BenchmarkAzureTestingManaged -benchtime=1x -cover -v
+          go test -tags testing_azure -bench=BenchmarkAzureTestingHA -benchtime=1x -cover -v
 
       - name: integration testing
         if: ${{ matrix.os != 'windows-latest' }}
