+++ .github/workflows/testingCivo.yml
@@ -80,8 +80,8 @@ jobs:
         working-directory: test
         if: ${{ matrix.os == 'windows-latest' }}
         run: |
-          go test -bench=BenchmarkCivoTestingManaged -benchtime=1x -cover -v
-          go test -bench=BenchmarkCivoTestingHA -benchtime=1x -cover -v
+          go test -tags testing_civo -bench=BenchmarkCivoTestingManaged -benchtime=1x -cover -v
+          go test -tags testing_civo -bench=BenchmarkCivoTestingHA -benchtime=1x -cover -v
 
       - name: integration testing
         if: ${{ matrix.os != 'windows-latest' }}
