+++ Makefile
@@ -167,32 +167,32 @@ unit_test_ksctl_stateimport: golang-test ## ksctl-stateimport unit test case
 mock_all: golang-test ## All Mock tests
 	@echo "Mock Test (integration)"
 	cd test/ && \
-		GOTEST_PALETTE="red,yellow,green" $(GO_TEST_COLOR) -bench=. -benchtime=1x -cover -v
+		GOTEST_PALETTE="red,yellow,green" $(GO_TEST_COLOR) -tags testing_aws,testing_civo,testing_azure -bench=. -benchtime=1x -cover -v
 
 .PHONY: mock-civo-ha
 mock_civo_ha: golang-test ## Civo HA mock test
 	cd test/ && \
- 		GOTEST_PALETTE="red,yellow,green" $(GO_TEST_COLOR) -bench=BenchmarkCivoTestingHA -benchtime=1x -cover -v
+ 		GOTEST_PALETTE="red,yellow,green" $(GO_TEST_COLOR) -tags testing_civo -bench=BenchmarkCivoTestingHA -benchtime=1x -cover -v
 
 .PHONY: mock-civo-managed
 mock_civo_managed: golang-test ## Civo managed mock test
 	cd test/ && \
- 		GOTEST_PALETTE="red,yellow,green" $(GO_TEST_COLOR) -bench=BenchmarkCivoTestingManaged -benchtime=1x -cover -v
+ 		GOTEST_PALETTE="red,yellow,green" $(GO_TEST_COLOR) -tags testing_civo -bench=BenchmarkCivoTestingManaged -benchtime=1x -cover -v
 
 .PHONY: mock-azure-managed
 mock_azure_managed: golang-test ## Azure managed mock test
 	cd test/ && \
- 		GOTEST_PALETTE="red,yellow,green" $(GO_TEST_COLOR) -bench=BenchmarkAzureTestingManaged -benchtime=1x -cover -v
+ 		GOTEST_PALETTE="red,yellow,green" $(GO_TEST_COLOR) -tags testing_azure -bench=BenchmarkAzureTestingManaged -benchtime=1x -cover -v
 
 .PHONY: mock-azure-ha
 mock_azure_ha: golang-test ## Azure HA mock test
 	cd test/ && \
- 		GOTEST_PALETTE="red,yellow,green" $(GO_TEST_COLOR) -bench=BenchmarkAzureTestingHA -benchtime=1x -cover -v
+ 		GOTEST_PALETTE="red,yellow,green" $(GO_TEST_COLOR) -tags testing_azure -bench=BenchmarkAzureTestingHA -benchtime=1x -cover -v
 
 .PHONY: mock-aws-ha
 mock_aws_ha: golang-test ## Aws HA mock test
 	cd test/ && \
- 		GOTEST_PALETTE="red,yellow,green" $(GO_TEST_COLOR) -bench=BenchmarkAwsTestingHA -benchtime=1x -cover -v
+ 		GOTEST_PALETTE="red,yellow,green" $(GO_TEST_COLOR) -tags testing_aws -bench=BenchmarkAwsTestingHA -benchtime=1x -cover -v
 
 
 .PHONY: mock-local-managed
