+++ pkg/controllers/managerClusterSelfManaged.go
@@ -65,18 +65,13 @@ func (manager *ManagerClusterSelfManaged) CreateCluster() error {
 		}
 	}()
 
-	fakeClient := false
-	if _, ok := helpers.IsContextPresent(controllerCtx, consts.KsctlTestFlagKey); ok {
-		fakeClient = true
-	}
-
 	if !helpers.ValidCNIPlugin(consts.KsctlValidCNIPlugin(client.Metadata.CNIPlugin)) {
 		err := log.NewError(controllerCtx, "invalid CNI plugin")
 		log.Error("handled error", "catch", err)
 		return err
 	}
 
-	if err := cloudController.InitCloud(client, stateDocument, consts.OperationCreate, fakeClient); err != nil {
+	if err := cloudController.InitCloud(client, stateDocument, consts.OperationCreate); err != nil {
 		log.Error("handled error", "catch", err)
 		return err
 	}
@@ -153,12 +148,7 @@ func (manager *ManagerClusterSelfManaged) DeleteCluster() error {
 		}
 	}()
 
-	fakeClient := false
-	if _, ok := helpers.IsContextPresent(controllerCtx, consts.KsctlTestFlagKey); ok {
-		fakeClient = true
-	}
-
-	if err := cloudController.InitCloud(client, stateDocument, consts.OperationDelete, fakeClient); err != nil {
+	if err := cloudController.InitCloud(client, stateDocument, consts.OperationDelete); err != nil {
 		log.Error("handled error", "catch", err)
 		return err
 	}
@@ -206,12 +196,7 @@ func (manager *ManagerClusterSelfManaged) AddWorkerPlaneNodes() error {
 		}
 	}()
 
-	fakeClient := false
-	if _, ok := helpers.IsContextPresent(controllerCtx, consts.KsctlTestFlagKey); ok {
-		fakeClient = true
-	}
-
-	if err := cloudController.InitCloud(client, stateDocument, consts.OperationGet, fakeClient); err != nil {
+	if err := cloudController.InitCloud(client, stateDocument, consts.OperationGet); err != nil {
 		log.Error("handled error", "catch", err)
 		return err
 	}
@@ -290,7 +275,7 @@ func (manager *ManagerClusterSelfManaged) DelWorkerPlaneNodes() error {
 		fakeClient = true
 	}
 
-	if err := cloudController.InitCloud(client, stateDocument, consts.OperationGet, fakeClient); err != nil {
+	if err := cloudController.InitCloud(client, stateDocument, consts.OperationGet); err != nil {
 		log.Error("handled error", "catch", err)
 		return err
 	}
